backend/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
python-multipart==0.0.20
numpy==2.1.3
opencv-python==4.11.0.86
pillow==11.1.0
mediapipe==0.10.33
transformers>=4.48.3
scikit-image==0.25.2
ultralytics>=8.0.0
scipy>=1.11.0
httpx>=0.27.0
python-dotenv>=1.0.0
